SectionRepository gets its very own file
[#85024596]
