+++ spec/lib/bookbinder/section_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../../lib/bookbinder/section'
+require_relative '../../../lib/bookbinder/repositories/section_repository'
 require_relative '../../helpers/tmp_dirs'
 require_relative '../../helpers/nil_logger'
 require_relative '../../helpers/spec_git_accessor'
@@ -9,7 +9,7 @@ module Bookbinder
 
     let(:logger) { NilLogger.new }
     let(:repository) {
-      SectionRepository.new(
+      Repositories::SectionRepository.new(
         logger,
         store: {},
         git_accessor: SpecGitAccessor
